requirements.txt
Flask
Flask-Session
Authlib
requests
dotenv
gevent
flask-sock
simple-websocket
gunicorn 
eventlet
flask-cors
pytest
Werkzeug
SQLAlchemy
Flask-SQLAlchemy
itsdangerous
psycopg2-binary
docker
